Move instance and induction scheme generation into separate modules #150
